refactor: update test suits with changes done by typescript transpiling
